Version 0.2.0 Pretty skeleton

+++ pland/src/lib/lobby.ts
@@ -10,9 +10,12 @@ import { browser } from '$app/env'
 import { checkPlants } from "./z3r/logic/Logic";
 
 export const Lobbies = new Map<string, Lobby>();
+let readOnly = false;
 
 export async function reloadLobbies() {
     if (Lobbies.size == 0) {
+        fs.access('lobbies', fs.constants.W_OK, () => { readOnly = true});
+        if (readOnly) return;
         if (fs.existsSync('lobbies')) {
             fs.readdirSync('lobbies').forEach(async lobbyFile => {
                 let file = fs.readFileSync(`lobbies/${lobbyFile}`).toString();
@@ -54,7 +57,7 @@ export interface ILobby {
 }
 
 function saveLobby(lobby: Lobby) {
-    if (!browser) {
+    if ((!browser) && (!readOnly)) {
         fs.writeFileSync(`lobbies/${lobby.lobby!.slug}`, JSON.stringify(lobby));
     }
 }
@@ -164,9 +167,7 @@ export default class Lobby {
     async checkAllReady() {
         if ((this.lobby.entrants.length == this.lobby.max_entrants) && (this.lobby?.entrants.every(entrant => entrant.ready))) {
             const allItemsPlanted = this.lobby.entrants.flatMap(entrant => entrant.plantedItems);
-            console.log(allItemsPlanted);
             const allLocationsPlanted = this.lobby.entrants.flatMap(entrant => entrant.plantedLocations);
-            console.log(allLocationsPlanted);
             const {plantable, messages} = checkPlants(this.world as World, allItemsPlanted, allLocationsPlanted);
             if (plantable) {
                 this.lobby.ready_to_roll = true;
